+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+ - Moved partition control open state and event handling into abstraction
+
+## [11.0.0] - 2019-02-11
 ### Added
  - Added ConferencePoint ids under Room in configuration